implemented undo/redo

+++ lab3/Console.c
@@ -7,34 +7,59 @@
 #include "Console.h"
 #include "FlieService.h"
 
-void UIAdd(Vector *repository, char *token);
-void UIDelete(Vector *repository, char *token);
-void UIUpdate(Vector *repository, char *token);
+bool UIAdd(Vector *repository, char *token);
+bool UIDelete(Vector *repository, char *token);
+bool UIUpdate(Vector *repository, char *token);
 void UIList(Vector *repository, char *token);
 void UIListAll(Vector *repository);
 void UIListOfType(Vector *repository, char *token);
+void UIListYear(Vector *repository, char *token);
 void printFileRepository(Vector *repository);
 
-void runConsole(Vector* fileRepository){
+void runConsole(Vector* fileRepository, Vector* repositoryHistory){
     char* command = (char*) malloc(sizeof(char) * 100);
     char* input = (char*) malloc(sizeof(char) * 100);
     char* token;
+    addToVector(repositoryHistory, fileRepository);
+    int currentIndexInHistory = 0;
     bool inputNotExit = true;
     while(inputNotExit){
         scanf(" %[^\n]s", input);
         token = strtok(input, " ");
         strcpy(command, token);
+        Vector* currentVector = repositoryHistory->copyFunction(repositoryHistory->buffer[currentIndexInHistory]);
         if(strcmp(command, "add") == 0){
-            UIAdd(fileRepository, token);
+            if(UIAdd(currentVector, token)){
+                removeAllFromIndex(repositoryHistory, currentIndexInHistory + 1);
+                addToVector(repositoryHistory, currentVector);
+                currentIndexInHistory = repositoryHistory->size - 1;
+            }
         } else if(strcmp(command, "delete") == 0){
-            UIDelete(fileRepository, token);
+            if(UIDelete(currentVector, token)){
+                removeAllFromIndex(repositoryHistory, currentIndexInHistory + 1);
+                addToVector(repositoryHistory, currentVector);
+                currentIndexInHistory = repositoryHistory->size - 1;
+            }
         } else if(strcmp(command, "update") == 0){
-            UIUpdate(fileRepository, token);
+            if(UIUpdate(currentVector, token)){
+                removeAllFromIndex(repositoryHistory, currentIndexInHistory + 1);
+                addToVector(repositoryHistory, currentVector);
+                currentIndexInHistory = repositoryHistory->size - 1;
+            }
         } else if(strcmp(command, "list") == 0){
-            UIList(fileRepository, token);
+            UIList(currentVector, token);
+        } else if(strcmp(command, "undo") == 0){
+            if(currentIndexInHistory > 0){
+                currentIndexInHistory--;
+            }
+        } else if(strcmp(command, "redo") == 0){
+            if(currentIndexInHistory < repositoryHistory->size - 1){
+                currentIndexInHistory++;
+            }
         } else if(strcmp(input, "exit") == 0){
             inputNotExit = false;
         }
+        destroyVector(currentVector);
     }
     FREE(command);
     FREE(input);
@@ -44,16 +69,17 @@ void UIList(Vector *repository, char *token){
     token = strtok(NULL, " ");
     if(token == NULL){
         UIListAll(repository);
+    } else if(strtol(token, NULL, 10)) {
+        UIListYear(repository, token);
     } else {
         UIListOfType(repository, token);
     }
 }
 
 void UIListOfType(Vector *repository, char *type){
-    Vector* filesOfType = getVectorOfType(repository, type);
+    Vector* filesOfType = getFilesOfType(repository, type);
     printFileRepository(filesOfType);
     destroyVector(filesOfType);
-
 }
 
 void UIListAll(Vector *repository){
@@ -68,7 +94,7 @@ void printFileRepository(Vector *repository){
     }
 }
 
-void UIUpdate(Vector *repository, char *token){
+bool UIUpdate(Vector *repository, char *token){
     int number, newYear;
     char* newType = (char*) malloc(sizeof(char) * 100);
     char* newState = (char*) malloc(sizeof(char) * 100);
@@ -84,23 +110,35 @@ void UIUpdate(Vector *repository, char *token){
 
     token = strtok(NULL, ", ");
     newYear = atoi(token);
+    if(hasFileWithCatalogNumber(repository, number)){
+        File *fileToAdd = createFile(number, newState, newType, newYear);
+        updateFile(repository, fileToAdd);
+        destroyFile(fileToAdd);
+        FREE(newType);
+        FREE(newState);
+        return true;
+    } else{
+        FREE(newType);
+        FREE(newState);
+        return false;
+    }
 
-    File* fileToAdd = createFile(number, newState, newType, newYear);
-    updateFile(repository, fileToAdd);
-    destroyFile(fileToAdd);
 
-    FREE(newType);
-    FREE(newState);
 }
 
-void UIDelete(Vector *repository, char *token){
+bool UIDelete(Vector *repository, char *token){
     int number;
     token = strtok(NULL, " ");
     number = atoi(token);
-    removeFile(repository, number);
+    if(hasFileWithCatalogNumber(repository, number)){
+        removeFile(repository, number);
+        return true;
+    } else{
+        return false;
+    }
 }
 
-void UIAdd(Vector *repository, char *token){
+bool UIAdd(Vector *repository, char *token){
     int newNumber, newYear;
     char* newType = (char*) malloc(sizeof(char) * 100);
     char* newState = (char*) malloc(sizeof(char) * 100);
@@ -116,11 +154,24 @@ void UIAdd(Vector *repository, char *token){
 
     token = strtok(NULL, ", ");
     newYear = atoi(token);
-    File* fileToAdd = createFile(newNumber, newState, newType, newYear);
-    addFile(repository, fileToAdd);
-    destroyFile(fileToAdd);
-    FREE(newType);
-    FREE(newState);
+    if(!hasFileWithCatalogNumber(repository, newNumber)){
+        File *fileToAdd = createFile(newNumber, newState, newType, newYear);
+        addFile(repository, fileToAdd);
+        destroyFile(fileToAdd);
+        FREE(newType);
+        FREE(newState);
+        return true;
+    } else{
+        FREE(newType);
+        FREE(newState);
+        return false;
+    }
+}
+
+void UIListYear(Vector *repository, char *token){
+    Vector* filesBeforeYear = getFilesBeforeYearSorted(repository, atoi(token));
+    printFileRepository(filesBeforeYear);
+    destroyVector(filesBeforeYear);
 }
